requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.2
frozenlist==1.4.1
future==1.0.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
line-bot-sdk==3.10.1
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
